+++ lib/src/utils/connectivity_logger.dart
@@ -3,10 +3,10 @@ import 'package:flutter/foundation.dart';
 /// A simple logger that works with WebAssembly by using Flutter's built-in
 /// debugPrint function instead of dart:io dependencies.
 class ConnectivityLogger {
-  static bool _enabled = kDebugMode;
+  bool _enabled = kDebugMode;
 
   /// Enable or disable logging
-  static void setEnabled(bool enabled) {
+  void setEnabled(bool enabled) {
     _enabled = enabled;
   }